Speed up end to end test
This is done by reducing the polling interval and number of
confirmations in the test. On my machine the test is nearly twice as
fast now.

I also added more explicit explanations why mining and waiting is
necessary.

A cleaner way to wait would be to wait until specific events come in,
like it is done in
https://github.com/raiden-network/raiden-contracts/blob/master/raiden_contracts/tests/test_monitoring_service.py .
Unfortunately, the `MockRaidenNode` methods used in this test don't
return tx hashes, which are required for the event handlers used in the
raiden_contracts test.