Keep RequestCollector running
Before, the RC exited immediately on startup. Now the process control is
given to `GMatrixClient.listen_forever`, which keeps the RC running until
the MatrixListener fails.

We could remove the `Greenlet` superclass from `RequestCollector` and
`MatrixListener`, as well as the `_run` and `stop` methods if we don't
plan to run them as greenlets at all. Currently we still do that in some
tests.