Enable pylint `protected-access` check
mostly by disabling it wherever it caused a problem. For tests, that is
OK, but we should improved the access to `_members` in the Matrix code.
Even if it's just by declaring that attribute as public.

Preferably I would disable this check for tests, but I didn't find a way
to do that without running pylint twice. And pylint is slow enough when
run once.