Allow running tests in parallel with pytest-xdist
The blocker for this was that gevent monkeypatching broke xdist. By
skipping `subprocess` and `thread` monkeypatching, this can be avoided.

We also need to make sure that not all processes choose the same port.
At least in my tests, choosing a free port by binding to port 0 did not
cause problematic race conditions, although I'm not sure that it always
works.