Configure CircleCI
In the raiden client we already use CircleCI and are happier with it
than with Travis. The change seems simple enough and even with this
basic configuration, we are faster than Travis (~2:45 vs ~3:45).

I'll keep both Travis and CircleCI on for a while to see if things run
as intended. We can remove Travis when we're confident enough. I just
disabled the codecov upload by Travis to avoid duplicate uploads.