Fix check for already closed channels
The problem was that we changed the `last_known_block` in the `Context`
from an independent value towards a a property returning the value of
`latest_known_block` of `BlockchainState`, which is only updated when
all events for a given block range are processed.

This led to the following condition to always be `False`.
```py
settle_period_end_block = event.block_number + channel.settle_timeout
    settle_period_over = settle_period_end_block < context.latest_confirmed_block
    if not settle_period_over:
       ...
```

This in turn let the MS to try to send an update to the blockchain for a
closed channel.

We have tests for that behavior, however we wrongly updated the
`BlockchainState` as well, which led to the tests not catching this.