Catch exceptions in event handlers
This makes it harder for attackers to take down a MS. Before, even
restarting would not have helped in many cases because the same Event
would have been retried forever. Now, events are discarded when they
throw an exception and the MS continues to run.

To make sure that we don't ignore real errors, the exceptions are logged
and sentry issues are created.

There are no tests for this because I did not see a simple enough way to
exercise this in a meaningful manner.