Simplify builder
The builder will now only care about the services, not about updating
itself, the docker-compose-files or traefik. These changes are rare and
can be done manually.

The builder (or the `update.sh` to be more precise) is also now able to
understand the difference between the `dev` and `test` deployments.