Don't replace gevent's error handler
If we use our own error handler, the processes are not cleaned up on
exception because we directly stop the process via `exit`, which raises
a `SystemExit` exceptions inside the hub. This is not allowed, as
http://www.gevent.org/api/gevent.hub.html#gevent.hub.Hub.handle_error
says:

> Applications can set a property on the hub instance with this same
signature to override the error handling provided by this class. This is
an advanced usage and requires great care. This function must not raise
any exceptions.

Alternative approach to error handling:
* If you spawn a greenlet you must check for errors using `.get` or
`joinall(..., raise_error=True)` to collect and raise the exceptions.
* For long running greenlets, regularly use `joinall` in combination with
a timeout.
* Once we have multiple child greenlets, we should kill all child
greenlets once one of them raises an exception. This way try/finally
clauses and context managers that are used in the greenlet have a chance
to do the cleanup. Right now, we don't have more than one child
greenlet, so I did not introduce code for this, yet.

If we fail to collect exceptions from a greenlet, the raised exception
is still printed, but not won't crash the process. Suggestions on how to make
that case more obvious are welcome!