Using persistent data for Path debug endpoint
If the PFS is restarted while a scenario is running the data of the PFS
requests will be lost, which has led to scenario failures. This commit
changes the debug endpoint to use data persisted in the database, so
that restarts will no lead to a problem.

Note: This is relying on the feedback data, which does not contain all
the requests. Specifically, requests which failed with a validation
error are not included in the table, therefore are still successtible to
loss of data.