Improve typing around creation of feedback tokens
In the review of this PR we noticed a bug that wasn't catched by the
tests. It turns out that the code worked by accident, because some types
differed from the annotations which mypy didn't find (because `Any` was
used).

This commit removes the use of `Any` and fixes the following fall-out.
With the changes the bug was catched by the tests.