Use `AsyncResult` to raise errors while waiting
We were waiting on `event`s that never came when the greenlet we're
waiting for threw an exception at an unfortunate time. Because we we're
stuck in the `wait`, we did not event collect and raise the exception
from the underlying greenlet.

If should probably have split some of the changes in `matrix.py` into a
separate commit, but I don't think it's worth the effort to pull the
changes apart, now.