Check block numbers instead of locking exclusively
In #5443, we had a problem with two PFSs sharing the same DB, which
lead to each of the PFSs processing roughly half of the blocks. Since
that was time consuming to debug, we should prevent that situation from
happening, again.

Our first approach was to prevent any access to the db by other
processes by acquiring an exclusive lock during the whole PFS runtime.
But this also prevents us from claiming PFS fees while the PFS is
running, since we need to mark the claimed IOUs in the db.

Using a single transaction when processing a block does not solve this
problems, since each PFS could still read the block number, process the
block and save the block number without any conflict. The processing of
blocks would only be serialized between the two PFSs, which does not
solve the problem.

The solution is to keep the last processed block in memory, so that
other processes writing to the db will not cause skipped blocks. To make
the bad situation of two PFSs on on db obvious, an assert is added that
compares the block number in memory with the one in the db.