Explicitly import fixtures from contracts
Without this, we get a mypy failure for a fixture that is redefined in
this repo. `import *` is a bad practice anyway, since it can lead to
failures at surprising times and makes it hard to trace where functions
actually come from.

A tricky thing about this PR was adding `auto_revert_chain`, since it is
not explicitly used by any test, but is required for tests to work
reliably.