Add `suggest_partner` api endpoint (#834)
When a Raiden node joins a token network, it should connect to nodes
which will be able to mediate payments. The PFS has a global view and
should have an easier time to find such nodes than the client.

See https://github.com/raiden-network/raiden-services/issues/684

The suggested partners are cached for one hour, which makes it
reasonable to offer this service without requiring fees. The amount of
results is fixed to make caching simpler.

The results are ordered and choosing just the first entry will be
sufficient for most use cases. For less common cases and for debugging,
additional scoring information is provided (The overall score as
`score`, as well as `centrality`, `uptime` and `capacity`). There is no
long term guarantee regarding the meaning of the specific values, but
greater values will always be better.
The current scoring is not very sophisticated, but should be good enough
as a starting point. Improvements to the scoring will be possible
without breaking compatibility.

I will start adding information about this endpoint to the spec once
this is reviewed and merged.