Refactor Prometheus RED metrics to raiden-libs
Since the MS as well as the PFS will share common metrics (especially Error and Processing time metrics),
the common functionality was refactored to the raiden-libs.
Error metrics will be available under the same namespaces and labels for all services
(e.g. "events_log_errors_total" with label "error_category":"state"),
so it is the responsibility of the Prometheus endpoint consumer to dispatch
the different endpoints for different services.