Change testing setup for metrics and get rid of teardown
The current teardown mechanism for the global state of the prometheus metrics
do not work in conjunction with metrics that query dynamically bound callback methods.
Since modifying the teardown mechanism would get very involved with replicating behaviour
from the prometheus_client library, the decision was made to accept global state in the tests
and only test for deltas on the metrics. In every test where a change of value in a metric
has to be asserted, the state of the metrics registry has to be saved before initiating
the action that will trigger the metric's value change.

Additionally, some convenience functionality is introduced, allowing for
easier retrieval of metrics objects, as well as label-dictionary construction.