Add PFS metrics
This adds basic metrics for the PathFindingService,
such as Events and Message handling RED metrics
as well as IOU's claimed metrics.
Note that for the latter,
every poll from a Prometheus server will trigger two requests
to the PFS database, where one of them will iterate over all
claimed IOUs and sum up the claimed token amount manually.
At the moment, token-amounts are stored in hex-format in the DB,
which is why it is not more efficient to do the summing on the SQLite
side.