Return PFS broadcast room id in info response
This is necessary after
- raiden-network/raiden-service-bundle#187
- raiden-network/raiden#6442

With those PRs the set of federation servers can become disjunct between
different Raiden client versions. Therefore we can't rely on the
previous check in the Raiden client if the PFS' server is in the known
servers set.

The RoomID is uniquely identifying a specific broadcasting room and can
therefore be used as a better information source for this check.