Use unstable contract env for goerli docker setup
Since https://github.com/raiden-network/raiden-services/pull/951, we can
use the unstable contracts environment. We want to do that for the
development machines that are run using the docker-compose. Since that
environment only exists for goerli, we only enable it for the goerli
services.

Relates to https://github.com/raiden-network/light-client/issues/2521