Cache UDC balances until BalanceReduced event
This reduces the number of RPC calls necessary to validate an IOU. If
the UDC balance has not been reduced since we last fetched the balance,
no RPC calls are necessary at all. This is important because it can add
a serious amount of latency to the route request.

TODO: fetch the events required for the invalidation