adapt comparison of event trigger timestamps
Events triggered for a specific timestamp must be compared based on the current
time itself and not the time of the latest block. The last block might be long
ago but the trigger timestamp has been passed.
The `latest_timestamp` property of the `Context` instance is still available. It
might become removed when all references got cleaned up.

This includes a small change to revert the tiny utility function on the
`Context` instance. It only gets used once and just doesn't fit into the
definition of what the context represents.
Furthermore some naming of the new timing terminology has been applied.