adapt IOU validation and claiming for timestamps
Timestamps are now claimable until a specific timestamp. This change has not yet
been reflected by the validation of received IOUs from clients nor for the
script to claim IOUs.
It has been agreed on that clients will per default send a new IOU that expires
within a month. Thereby it is totally fine that the pathfinding service expects
an minimum expiration date of one week in the future. This margin is by far
enough to also compensate the delay between the point in time the client
calculates the `claimable_until` property and the point in time the PFS receives
the message and validates it. In the past both peers have used the last seen
block as reference point (though this also must not have been the same all the
time).