update eth-tester chain identifier
There is still the issue that web3 and eth-tester are incompatible when it comes
to the chain identifier. The issue opened by a colleague is still open. The
workaround was to have a test constant that mirrors the identifier as defined by
eth-tester and use it instead of asking web3. This constant got out-of-date with
the recent update of eth-tester.
This was causing issues with invalid signatures for balance proofs between test
code and contract code in the EVM.