improve naming for monitor event handling
This change got triggered by the issue that with the recent changes we ended up
with a wrong logging message as of `settle_until=settleable_after`. Anyway, this
gave us the hint that the naming in the event handling of a closed channel is
not fully right yet. The core if that is that from the old code the period of
time where a channel got closed to the point it can be settled was named the
`settle_period`. This is simply wrong as it is basically the opposite as the
channel can **not** be settled. The actually interesting part here is if the
monitoring service can hand in an updated balance proof for the non-closing
channel participant or not. Based on that the variables and logging message got
adapted.