fix usage of eth-tester import in production
Importing anything from eth-tester does not work in production as it is a test
dependency that is not available. In result the monitoring service was failing
to start.
The import was used to test if a failed update balance proof transaction should
be retried or not. The challenge is that in test mode, web3 forwards eth-tester
exceptions untouched. That is not the same in production mode. Thereby the error
looks slightly different. But both are based on the BaseException class. So it
is possible to just verify their error message without matching the exact error
class.

This bug got introduced with a recent feature implementation. Unfortunately the
CI pipeline was not properly working, which allowed to introduce this mistake.