automatically set gas price strategy for ms
This change makes the gas price parameter for the monitoring service optional.
In case it is not manually set by the user, it will be defined automatically. In
this case it checks if the connected network is based on Arbitrum or not. In the
former case it now defaults to the `rpc` based strategy, where the latter case
is the former click option default.