retry too early claim reward transactions
As for the monitor transaction, the claim reward transaction can be send too
early due to unsynchronized clocks between the monitoring service and the
blockchain node. Therefore it is necessary to catch these special cases and put
the event (that triggers the transaction) into the retry mechanism that got
already established.