fix wrong UTC timestamp usage
Using the `datetime.utcnow()` function creates a timezone unaware `datetime`
instance. If now the `timestamp()` function gets called on such an instance, it
will become "converted" to the locale timezone again. That is not an issue as
long as we use it as pure `datetime` object. But at the locations where we need
to format into to POSIX timestamp, we run into problems with wrong values.
The solution to this is to create a timezone **aware** `datetime` instance. As
this results into more code than before, it got exported into a new utility
function which provides some convenience everywhere used.